Reduce dependencies of the core package.
Separate a base package that does not have so many
dependencies; especially on iolib, since it appears
to have very limited OS portability.

Also note that the reader case macro does not quite
work if the body refers to symbols in non-existant
packages, hence the addition of an ordinary guard.